// src/client/src/app/modules/admin/identity/models/user.ts
export interface User {
    id: string;
    userName: string;
    firstName: string;
    lastName: string;
    email: string;
    emailConfirmed: boolean;
    curentPassword: string;
    password: string;
    confirmPassword: string;
    isActive: boolean;
    phoneNumber: string;
    phoneNumberConfirmed: boolean;
    profilePictureUrl: string;
    employeeId: string;
}